!O (Sandbox, Flowgraph) Cleanup FG Tree Reload on Object Change (part 1). Improves level loading and object editing (Approved by samuelk) Review: ID 66083 by Author: Ines Almeida, Reviewer: Michael Madsen, Observer: Sebastien Laurent, Observer: Nico Moss, Observer: Samuel Kahn
Copied from Perforce
 Change: 1516245

+++ Code/CryEngine/CryAction/MaterialEffects/MaterialEffects.cpp
@@ -906,7 +906,7 @@ void CMaterialEffects::ReloadMatFXFlowGraphs()
 	m_pMaterialFGManager->ReloadFlowGraphs();
 }
 
-int CMaterialEffects::GetMatFXFlowGraphCount() const
+size_t CMaterialEffects::GetMatFXFlowGraphCount() const
 {
 	return m_pMaterialFGManager->GetFlowGraphCount();
 }
